Remove PhysicalClient Resolver
