+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoRaWan.NetworkServer/Common/PhysicalClient.cs
@@ -6,7 +6,14 @@ namespace LoRaWan.NetworkServer.Common
     using System;
     using System.Collections.Generic;
     using System.Text;
+    using BasicStation;
+    using LoRaTools.ADR;
+    using LoRaWan.NetworkServer.ADR;
+    using LoRaWan.NetworkServer.PacketForwarder;
+    using LoRaWan.Shared;
     using Microsoft.Azure.Devices.Client;
+    using Microsoft.Extensions.Caching.Memory;
+    using Microsoft.Extensions.Logging;
 
     public class PhysicalClient
     {
@@ -29,6 +36,43 @@ namespace LoRaWan.NetworkServer.Common
             this.messageDispatcher = messageDispatcher;
             this.loRaDeviceAPIService = loRaDeviceAPIService;
             this.loRaDeviceRegistry = loRaDeviceRegistry;
+        }
+
+        /// <summary>
+        /// Factory method to instantiate the correct UDP endpoint.
+        /// </summary>
+        public static IPhysicalClient Create()
+        {
+            var configuration = NetworkServerConfiguration.CreateFromEnviromentVariables();
+
+            if (configuration.UseBasicStation)
+            {
+                Logger.Log("Using physical client with basic station implementation", LogLevel.Information);
+                return new BasicStation();
+            }
+            else
+            {
+                Logger.Log("Using physical client with the packet forwarder implementation", LogLevel.Information);
+                var loRaDeviceAPIService = new LoRaDeviceAPIService(configuration, new ServiceFacadeHttpClientProvider(configuration, ApiVersion.LatestVersion));
+                var frameCounterStrategyProvider = new LoRaDeviceFrameCounterUpdateStrategyProvider(configuration.GatewayID, loRaDeviceAPIService);
+                var deduplicationStrategyFactory = new DeduplicationStrategyFactory(loRaDeviceAPIService);
+                var adrStrategyProvider = new LoRaADRStrategyProvider();
+                var cache = new MemoryCache(new MemoryCacheOptions());
+                var dataHandlerImplementation = new DefaultLoRaDataRequestHandler(configuration, frameCounterStrategyProvider, new LoRaPayloadDecoder(), deduplicationStrategyFactory, adrStrategyProvider, new LoRAADRManagerFactory(loRaDeviceAPIService), new FunctionBundlerProvider(loRaDeviceAPIService));
+                var connectionManager = new LoRaDeviceClientConnectionManager(cache);
+                var loRaDeviceFactory = new LoRaDeviceFactory(configuration, dataHandlerImplementation, connectionManager);
+                var loRaDeviceRegistry = new LoRaDeviceRegistry(configuration, cache, loRaDeviceAPIService, loRaDeviceFactory);
+
+                var messageDispatcher = new MessageDispatcher(configuration, loRaDeviceRegistry, frameCounterStrategyProvider);
+                var udpServer = new UdpServer(configuration, messageDispatcher, loRaDeviceAPIService, loRaDeviceRegistry);
+
+                // TODO: review dependencies
+                var classCMessageSender = new DefaultClassCDevicesMessageSender(configuration, loRaDeviceRegistry, udpServer, frameCounterStrategyProvider);
+                dataHandlerImplementation.SetClassCMessageSender(classCMessageSender);
+
+                udpServer.SetClassCMessageSender(classCMessageSender);
+                return udpServer;
+            }
         }
     }
 }
